backend/requirements.txt
dj-database-url==0.5.0
Django==2.0.6
django-configurations==2.0
django-cors-headers==2.2.1
django-templated-mail==1.1.1
djangorestframework==3.8.2
djoser==1.1.5
psycopg2-binary
pytz==2018.4
uwsgi==2.0.17
django-celery-beat==1.5
celery==4.2.1
django-celery-results==1.0.1
django-redis==4.7.0
requests==2.13.0
pyamqp
jupyter
pandas
numpy
django-extensions